Separate TrimAudioSliderView in its own class
